feat(abi): introduce macro impl_codec for encoding ABIs
